Clear region if previously defined
Say we have a thin film where m = Uniform(0, 0, 1), and we define a
circular region 1 with m = Uniform (0, 0, -1). Let's say we now want
region 1 to be centred at some new coordinates (x', y'). We "reset" the
system by calling m = Uniform(0, 0, 1) again, then redefine the region
by again calling DefRegion(1, Circle(...).Transl(...)), and set m to be
Uniform (0, 0, -1) in the region. Currently, both circles at (x, y) and
(x', y') will be seen in the output, as region 1 wasn't reset back to
zero everywhere. This commit defines a new function clear(), and calls
it when DefRegion(id, shape) is called.

I ran into this problem when I was running a simulation in which I was
using a for loop to loop through various positions of a specific
texture, and output the energy of each configuration.